requirements.txt
requests
python-dotenv
playwright
pydantic
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
